+++ CHANGELOG.md
@@ -11,7 +11,23 @@ Please make sure you have a backup (preferably in git) of your code before runni
 - Add new option [issue #31](https://github.com/gmlewis/flutter-stylizer/issues/31):
   - `groupAndSortVariableTypes` (default: `false`)
 
-## [0.1.1] - 2021-04-26
+## [v0.1.7] - 2022-01-12
+
+- Fix [issue #26](https://github.com/gmlewis/flutter-stylizer/issues/26) caused by `Function()`.
+
+## [v0.1.6] - 2022-01-08
+
+- Process all `mixin` blocks in addition to all `class` blocks.
+
+## [v0.1.5] - 2021-12-12
+
+- `private-other-methods` can be added to the member ordering.
+
+## [v0.1.4] - 2021-11-03
+
+- Add "exclude" file globbing in the config file.
+
+## [v0.1.1] - 2021-04-26
 
 - Initial release. This version was successfully run on the Flutter code base
   and should be ready for use.